feat: support generate static html
