+++ examples/with-prerender-mpa/package.json
@@ -2,18 +2,16 @@
   "name": "exmaple-with-prerender-mpa",
   "description": "",
   "dependencies": {
-    "ice.js": "^1.0.11",
     "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "build-plugin-prerender": "^1.5.1"
+    "react-dom": "^16.4.1"
   },
   "devDependencies": {
     "@types/react": "^16.9.13",
     "@types/react-dom": "^16.9.4"
   },
   "scripts": {
-    "start": "icejs start",
-    "build": "icejs build"
+    "start": "../../packages/icejs/bin/ice-cli.js start --mode dev",
+    "build": "../../packages/icejs/bin/ice-cli.js build --mode prod"
   },
   "engines": {
     "node": ">=8.0.0"
