+++ examples/with-prerender-spa/build.json
@@ -1,17 +1,1 @@
-{
-  "plugins": [
-    [
-      "build-plugin-prerender",
-      {
-        "routes": [
-          "/",
-          "/about",
-          "/dashboard"
-        ],
-        "renderer": {
-          "headless": false
-        }
-      }
-    ]
-  ]
-}
+{}
