+++ examples/with-prerender-spa/src/pages/Home/index.tsx
@@ -14,7 +14,7 @@ const Home = () => {
             <Link to="/Dashboard">Dashboard</Link>
           </li>
           <li>
-            <Link to="/about">About</Link>
+            <Link to="/about/1">About</Link>
           </li>
         </ul>
       </div>
