+++ packages/icejs/src/getBuiltInPlugins.ts
@@ -37,12 +37,10 @@ const getBuiltInPlugins: IGetBuiltInPlugins = (userConfig) => {
     'build-plugin-ice-config',
     'build-plugin-ice-mpa',
     'build-plugin-ice-request',
-    'build-plugin-helmet'
+    'build-plugin-helmet',
+    // server render
+    'build-plugin-ice-ssr'
   ];
-
-  if (userConfig.ssr) {
-    plugins.push('build-plugin-ice-ssr');
-  }
 
   // add store plugin
   if (!Object.prototype.hasOwnProperty.call(userConfig, 'store') || userConfig.store !== false) {
